% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokehelperr.R
\name{calc_balance}
\alias{calc_balance}
\title{calc_balance}
\usage{
calc_balance(resistances, weaknesses)
}
\arguments{
\item{resistances:}{list of resistances obtained from calc_resistances}

\item{weaknesses:}{list of weaknesses obtained from calc_weaknesses}
}
\value{
balance : float
measure of how balanced the team is.
}
\description{
Calculate a measure of how balanced a team is using its
weaknesses and resistances.
Higher values indicate a more balanced team.
}
\examples{
calc_balance(c(Normal = 0, Fire = 3), c(Normal = 0, Fire = 3))

}
