% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokehelperr.R
\name{calc_resistances}
\alias{calc_resistances}
\title{calc_resistances}
\usage{
calc_resistances(team_types)
}
\arguments{
\item{team_types}{a list of Pokemon types.}
}
\value{
resistance_list a list of all Pokemon names and integers
team has to that type as values.
}
\description{
Given a list of pokémon types present on a player's team,
calculate a measure of how resistant the team is to each type in the game.

Creates a list in which the names are each of the 18 types
in the game, and the values are integers measuring the level of
resistance the input team has to that type. Higher values indicate a
higher level of resistance to that type (name).
}
\examples{
calc_resistances(list(c("Electric"), c("Fire", "Flying")))

}
