% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokehelperr.R
\name{calc_weaknesses}
\alias{calc_weaknesses}
\title{calc_weaknesses}
\usage{
calc_weaknesses(team_types)
}
\arguments{
\item{team_list}{: list of list of strings
list of pokémon types associated to the user's team obtained
via \code{get_types}}
}
\value{
weaknesses : list
a list containing all 18 pokémon types
and integers measuring the level of weakness the input team
has to that type as values.
}
\description{
Given a list of pokémon types present on a player's team,
calculate a measure of how weak the team is to each type in the game.

Creates a list in which the keys are each of the 18 types
in the game, and the values are integers measuring the level of
weakness the input team has to that key (type). Higher values indicate a
higher level of weakness to that type.
}
\examples{
calc_weaknesses(list(list("Electric"), list("Fire", "Flying")))

}
