% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokehelperr.R
\name{recommend}
\alias{recommend}
\title{recommend}
\usage{
recommend(current_team)
}
\arguments{
\item{current_team:}{list of strings
list of up to 5 pokémon names}
}
\value{
recommendation: string
the name of a pokémon that could be added to the input
team to make its weaknesses and resistances more balanced.
}
\description{
Given a team of up to 5 pokémon, recommend a
pokémon that could be added to the
current team to make its weaknesses and
resistances more balanced.

This function first determines which types the
team is most weak to and which types the team is
most resistant to via \code{calc_resistances} and
\code{calc_weaknesses}, and then makes its recommendation
based on this information.
}
\examples{
recommend(list('Pikachu', 'Eevee', 'Charizard'))

}
